requirements.txt
Flask
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
boto3
python-dotenv
langchain
langchain-google-genai

